fix: preserve entity API boundaries
