+++ src/entities/card/index.ts
@@ -1,6 +1,8 @@
 export { fetchCards, subscribeCards } from "./api/firestore";
 /** @public Separated read operations for the cross-Entity Card document boundary. */
 export { fetchCardReads, subscribeCardReads } from "./api/firestore";
+/** @public Separated read contract for the cross-Entity Card document boundary. */
+export type { CardRead } from "./api/firestore";
 export { generateCardId } from "./api/id";
 export { deleteCard, editCard, mutateCards } from "./api/mutations";
 export { useCard, useCards, useCardsByDeckId } from "./model/hooks";
@@ -12,6 +14,7 @@ export type {
   CardId,
   CardMutation,
   CardRaw,
+  RemoteCard,
 } from "./model/types";
 export {
   countCardsByDeckId,
