+++ src/entities/card/model/schema.ts
@@ -1,6 +1,5 @@
 import { z } from "zod";
 
-import { firestoreTimestampDateSchema } from "@/shared/api";
 import { isNonBlank } from "@/shared/lib/isNonBlank";
 
 const authenticatedUidSchema = z.string().min(1, "A confirmed user is required for remote Card writes");
@@ -47,29 +46,6 @@ export const cardSchema = cardCreateSchema.extend({
 export const localCardSchema = localCardCreateSchema.extend({
   createdAt: z.number(),
   updatedAt: z.number(),
-});
-
-// Reads accept compatible legacy values; command schemas enforce the stricter invariants required for new writes.
-export const cardDocumentSchema = z.object({
-  // Older documents may duplicate the Firestore document id in their stored fields.
-  id: z.string().optional(),
-  frontText: z.string(),
-  backText: z.string(),
-  tags: z.array(z.string()),
-  uniqueKey: z.string(),
-  deckId: z.string(),
-  uid: z.string(),
-  createdAt: z.number(),
-  updatedAt: z.number(),
-  deletedAt: z.number().nullable(),
-  score: z.number(),
-  numberOfSeen: z.number(),
-  lastSeenAt: z.number().optional(),
-  nextSeeingAt: firestoreTimestampDateSchema.optional(),
-  interval: z.number().optional(),
-  url: z.string().optional(),
-  startLine: z.number().optional(),
-  endLine: z.number().optional(),
 });
 
 // Zustand JSON storage serializes Dates as strings; hydration accepts only strings that restore to valid Dates.
