+++ src/entities/deck/model/schema.ts
@@ -49,25 +49,6 @@ export const localDeckSchema = localDeckCreateSchema.extend({
 
 export const persistedDeckStateSchema = z.object({ localDecks: z.array(localDeckSchema) });
 
-export const deckDocumentSchema = z.object({
-  // Older documents duplicate the Firestore document id in their data.
-  id: z.string().optional(),
-  // Read validation remains permissive for legacy data; command schemas enforce current write constraints.
-  name: z.string(),
-  url: z.string().optional(),
-  isPublic: z.boolean(),
-  uid: z.string(),
-  createdAt: z.number(),
-  updatedAt: z.number(),
-  deletedAt: z.number().nullable(),
-  scoreMax: z.number().nullable(),
-  scoreMin: z.number().nullable(),
-  selectedTags: z.array(z.string()),
-  tagAndFilter: z.boolean(),
-  category: z.string(),
-  convertToBr: z.boolean(),
-});
-
 export const deckEditSchema = editableDeckFieldsSchema.partial().extend({
   id: deckIdSchema,
   url: editableDeckFieldsSchema.shape.url.nullable(),
