+++ src/entities/deck/model/types.ts
@@ -1,7 +1,6 @@
 import type { z } from "zod";
 
 import type {
-  deckDocumentSchema,
   deckCreateSchema,
   deckEditSchema,
   deleteDeckSchema,
@@ -57,7 +56,14 @@ export type DeckView = DeckDomain & { localMode: boolean };
 /** Public Deck read model exposed outside the Entity. */
 export type Deck = DeckView;
 /** Field shape stored in one remote Deck Firestore document. */
-export type DeckDocument = z.infer<typeof deckDocumentSchema>;
+export type DeckDocument = Omit<DeckDomain, "id" | "url"> & {
+  /** Legacy documents may duplicate the Firestore document id in their stored fields. */
+  id?: DeckId | undefined;
+  /** Boundary validation may preserve an explicit undefined before the DTO mapper removes it. */
+  url?: string | undefined;
+  uid: string;
+  deletedAt: number | null;
+};
 /** Validated payload used to create a remote Deck document. */
 export type DeckCreate = z.infer<typeof deckCreateSchema>;
 /** Input accepted at the remote Deck creation boundary. */
